requirements.txt
appdirs==1.4.4
argcomplete==1.12.2
click==7.1.2
Flask==1.1.2
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
peewee==3.14.0
Pillow==8.1.0
pkg-resources==0.0.0
pyserial==3.5
python-escpos==2.2.0
pyusb==1.1.1
PyYAML==5.4.1
qrcode==6.1
six==1.15.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.0
